Rame

+++ src/main/java/com/example/payment/dto/request/PaymentRequestDTO.java
@@ -16,7 +16,7 @@ public class PaymentRequestDTO {
     // Node.js에서 보내는 JSON의 Key 값과 변수명이 정확히 일치해야 자동 매핑.    
     private String orderId; // res.reservations의 reservation_id (결제 원장의 reference_id로 사용)
     private Long memberId;        // 결제자 member_id (지갑 조회를 위해 필수)
-    private Integer amount;       // 총 결제 금액 (total_price)
+    private Long amount;       // 총 결제 금액 (total_price)
     private String type;   // 요청 타입
     private String eventTitle;    // 결제 상세 내역에 기록할 공연명
